Remove old cameras and unecessary xmls
